+++ GameEngine/GameEngine/Engine/Systems/TriggerSystem.hpp
@@ -2,7 +2,7 @@
 
 #include <unordered_set>
 
-#include <Systems/System.h>
+#include <System/System.h>
 #include <Physics/TriggerListener.hpp>
 #include <Components/Collider.hpp>
 #include <Physics/ITriggerListener.hpp>
