+++ Reader/Systems/LifetimeSystem.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <Systems/System.h>
+#include <System/System.h>
 
 namespace AGE
 {
